# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
  - application/excel
  - application/vnd.ms-excel
  - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
paths:
  /agents/submitted/survey:
    x-swagger-router-controller: agents
    get:
      description: Returns total list of agents who submitted survey
      # used as the method name of the controller
      operationId: agentsSubmittedSurvey
      parameters:
        - $ref: "#/parameters/StartDate"
        - $ref: "#/parameters/EndDate"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /survey/submitted:
    x-swagger-router-controller: survey
    get:
      description: Returns total list of agents who submitted survey
      # used as the method name of the controller
      operationId: surveySubmitted
      parameters:
        - $ref: "#/parameters/StartDate"
        - $ref: "#/parameters/EndDate"
        - $ref: "#/parameters/DocType"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /survey/question:
    x-swagger-router-controller: survey
    get:
      description: Returns total list of agents who submitted survey
      # used as the method name of the controller
      operationId: surveyQuestion
      parameters:
        - $ref: "#/parameters/StartDate"
        - $ref: "#/parameters/EndDate"
        - $ref: "#/parameters/DocTypeNotReq"
        - $ref: "#/parameters/Question"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /survey/response:
    x-swagger-router-controller: survey
    get:
      description: Returns the survey responses
      operationId: surveyResponse
      parameters:
        - $ref: "#/parameters/SubsDTStart"
        - $ref: "#/parameters/SubsDTEnd"
        - $ref: "#/parameters/RespDTStart"
        - $ref: "#/parameters/RespDTEnd"
        - $ref: "#/parameters/DocTypeNotReq"
        - $ref: "#/parameters/Username"
        - $ref: "#/parameters/PageNo"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /survey/status/{status}:
    x-swagger-router-controller: actions
    post:
      description: Approve selected survey responses
      operationId: setStatus
      parameters:
        - $ref: "#/parameters/Status"
        - in: body
          name: selectedSurvey
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /survey/download/selected:
    x-swagger-router-controller: download
    get:
      description: Download selected survey responses
      operationId: downloadSelected
      parameters:
        - $ref: "#/parameters/SelectedSurvey"
      produces:
        - application/excel
      responses:
        200:
          description: Success
          headers:
            Content-type:
              type: string
              description: application/excel
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /survey/download/all:
    x-swagger-router-controller: download
    get:
      description: Download selected survey responses
      operationId: downloadAll
      parameters:
        - $ref: "#/parameters/SubsDTStart"
        - $ref: "#/parameters/SubsDTEnd"
        - $ref: "#/parameters/RespDTStart"
        - $ref: "#/parameters/RespDTEnd"
        - $ref: "#/parameters/DocTypeNotReq"
        - $ref: "#/parameters/Username"
        - $ref: "#/parameters/PageNo"
      produces:
        - application/excel
      responses:
        200:
          description: Success
          schema: 
            type: file
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user:
    x-swagger-router-controller: user
    post:
      description: Add user
      operationId: addUser
      parameters:
        - $ref: "#/parameters/UserData"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put: 
      description: Update user
      operationId: updateUser
      parameters:
        - $ref: "#/parameters/UserData"
        - $ref: "#/parameters/Key"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Delete user
      operationId: deleteUser
      parameters:
        - $ref: "#/parameters/Key"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: List all users
      operationId: listUsers
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  SuccessResponse:
    required:
      - status
      - message
    properties:
      status:
        type: integer
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string

parameters:
  StartDate:
    name: dateStart
    in: query
    description: The start date when agents submitted survey
    required: true
    type: string
  EndDate:
    name: dateEnd
    in: query
    description: The end date when agents submitted survey
    required: true
    type: string
  DocType:
    name: docType
    in: query
    description: The type of survey to search
    required: true
    type: string
    enum: [preEvent, postEvent]
  DocTypeNotReq:
    name: docType
    in: query
    description: The type of survey to search
    required: false
    type: string
    enum: [preEvent, postEvent]
  Question:
    name: question
    in: query
    description: The survey question
    required: true
    type: string
  SubsDTStart:
    name: subsDTStart
    in: query
    description: The start date when agents submitted survey
    required: false
    type: string
  SubsDTEnd:
    name: subsDTEnd
    in: query
    description: The end date when agents submitted survey
    required: false
    type: string
  RespDTStart:
    name: respDTStart
    in: query
    description: The start date when respondent response to survey question
    required: false
    type: string
  RespDTEnd:
    name: respDTEnd
    in: query
    description: The end date when respondent response to survey question
    required: false
    type: string
  Username:
    name: username
    in: query
    description: username of the one who submitted the survey
    required: false
    type: string
  PageNo:
    name: page
    in: query
    description: the page number
    required: false
    type: number
  Status:
    name: status
    in: path
    description: the status to be set
    required: true
    type: string
    enum:
      - APPROVED
      - REJECTED
  SelectedSurvey:
    name: selectedSurvey
    in: query
    required: true
    type: array
    items:
      type: string
  UserData:
    name: userdata
    in: body
    required: true
    schema:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        username:
          type: string
        password:
          type: string
  Key:
    name: key
    in: query
    required: true
    type: string
    
  